requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.4.2
PyPDF2==3.0.1
Pillow==10.0.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.23.3
cloudinary
streamlit
requests